PassportProviderService merges into ProviderService
