+++ service/src/main/java/bio/terra/externalcreds/services/ProviderService.java
@@ -1,6 +1,7 @@
 package bio.terra.externalcreds.services;
 
 import bio.terra.common.exception.BadRequestException;
+import bio.terra.common.exception.ForbiddenException;
 import bio.terra.common.exception.NotFoundException;
 import bio.terra.externalcreds.ExternalCredsException;
 import bio.terra.externalcreds.auditLogging.AuditLogEvent;
@@ -13,7 +14,10 @@ import bio.terra.externalcreds.models.CannotDecodeOAuth2State;
 import bio.terra.externalcreds.models.LinkedAccount;
 import bio.terra.externalcreds.models.LinkedAccountWithPassportAndVisas;
 import bio.terra.externalcreds.models.OAuth2State;
+import bio.terra.externalcreds.models.VisaVerificationDetails;
+import bio.terra.externalcreds.util.ProviderUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.annotations.VisibleForTesting;
 import java.security.SecureRandom;
 import java.sql.Timestamp;
 import java.time.Duration;
@@ -21,17 +25,21 @@ import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.core.OAuth2AuthorizationException;
 import org.springframework.security.oauth2.core.OAuth2ErrorCodes;
+import org.springframework.security.oauth2.core.OAuth2RefreshToken;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -45,6 +53,9 @@ public class ProviderService {
   public final ProviderTokenClientCache providerTokenClientCache;
   public final OAuth2Service oAuth2Service;
   public final LinkedAccountService linkedAccountService;
+  private final PassportService passportService;
+  private final JwtUtils jwtUtils;
+  private final AccessTokenCacheService accessTokenCacheService;
 
   public final AuditLogger auditLogger;
   public final SecureRandom secureRandom = new SecureRandom();
@@ -71,7 +82,10 @@ public class ProviderService {
       OAuth2Service oAuth2Service,
       LinkedAccountService linkedAccountService,
       AuditLogger auditLogger,
-      ObjectMapper objectMapper) {
+      ObjectMapper objectMapper,
+      PassportService passportService,
+      JwtUtils jwtUtils,
+      AccessTokenCacheService accessTokenCacheService) {
     this.externalCredsConfig = externalCredsConfig;
     this.providerOAuthClientCache = providerOAuthClientCache;
     this.providerTokenClientCache = providerTokenClientCache;
@@ -79,6 +93,9 @@ public class ProviderService {
     this.linkedAccountService = linkedAccountService;
     this.auditLogger = auditLogger;
     this.objectMapper = objectMapper;
+    this.passportService = passportService;
+    this.jwtUtils = jwtUtils;
+    this.accessTokenCacheService = accessTokenCacheService;
   }
 
   public Set<String> getProviderList() {
@@ -273,5 +290,276 @@ public class ProviderService {
         linkedAccount.getUserId(),
         linkedAccount.getProvider().toString(),
         responseBody);
+  }
+
+  public LinkedAccountWithPassportAndVisas createLink(
+      Provider provider,
+      String userId,
+      String authorizationCode,
+      String encodedState,
+      AuditLogEvent.Builder auditLogEventBuilder) {
+
+    var oAuth2State = validateOAuth2State(provider, userId, encodedState);
+    var providerClient = providerOAuthClientCache.getProviderClient(provider);
+    var providerInfo = externalCredsConfig.getProviderProperties(provider);
+    try {
+      var linkedAccount =
+          createLinkedAccount(
+              provider,
+              userId,
+              authorizationCode,
+              oAuth2State.getRedirectUri(),
+              new HashSet<>(providerInfo.getScopes()),
+              encodedState,
+              providerClient);
+
+      var linkedAccountWithPassportAndVisas = upsertLinkedAccount(providerInfo, linkedAccount);
+
+      logLinkCreation(Optional.of(linkedAccountWithPassportAndVisas), auditLogEventBuilder);
+      return linkedAccountWithPassportAndVisas;
+    } catch (OAuth2AuthorizationException oauthEx) {
+      logLinkCreation(Optional.empty(), auditLogEventBuilder);
+      throw new BadRequestException(oauthEx);
+    }
+  }
+
+  private LinkedAccountWithPassportAndVisas upsertLinkedAccount(
+      ProviderProperties providerInfo, ImmutablePair<LinkedAccount, OAuth2User> linkedAccount) {
+    if (ProviderUtils.isPassportProvider(providerInfo)) {
+      return linkedAccountService.upsertLinkedAccountWithPassportAndVisas(
+          jwtUtils.enrichAccountWithPassportAndVisas(
+              linkedAccount.getLeft(), linkedAccount.getRight()));
+    } else {
+      linkedAccountService.upsertLinkedAccount(linkedAccount.getLeft());
+      // returns a LinkedAccountWithPassportAndVisas with empty passports and visas
+      return new LinkedAccountWithPassportAndVisas.Builder()
+          .linkedAccount(linkedAccount.getLeft())
+          .build();
+    }
+  }
+
+  /**
+   * It is important to invalidate an expired link when it has a passport because downstream systems
+   * need to revoke authorizations that it grants. We don't need to actively do this when there are
+   * no passports, we just stop serving access token requests.
+   *
+   * @return the number of links invalidated
+   */
+  public int invalidateExpiredLinkedAccountsWithPassports() {
+    var expiredLinkedAccountsWithPassports =
+        linkedAccountService.getExpiredLinkedAccountsWithPassports();
+    expiredLinkedAccountsWithPassports.forEach(this::invalidateLinkedAccountAndRemovePassport);
+    return expiredLinkedAccountsWithPassports.size();
+  }
+
+  @VisibleForTesting
+  public void logLinkCreation(
+      Optional<LinkedAccountWithPassportAndVisas> linkedAccountWithPassportAndVisas,
+      AuditLogEvent.Builder auditLogEventBuilder) {
+    var passport =
+        linkedAccountWithPassportAndVisas.flatMap(LinkedAccountWithPassportAndVisas::getPassport);
+    var transactionClaim = passport.flatMap(p -> jwtUtils.getJwtTransactionClaim(p.getJwt()));
+    auditLogger.logEvent(
+        auditLogEventBuilder
+            .externalUserId(
+                linkedAccountWithPassportAndVisas.map(
+                    l -> l.getLinkedAccount().getExternalUserId()))
+            .auditLogEventType(
+                linkedAccountWithPassportAndVisas
+                    .map(x -> AuditLogEventType.LinkCreated)
+                    .orElse(AuditLogEventType.LinkCreationFailed))
+            .transactionClaim(transactionClaim)
+            .build());
+  }
+
+  public int validateAccessTokenVisas() {
+    var visaDetailsList = passportService.getUnvalidatedAccessTokenVisaDetails();
+
+    var linkedAccountIdsToRefresh =
+        visaDetailsList.stream()
+            .flatMap(
+                visaDetails ->
+                    validateVisaWithProvider(visaDetails)
+                        ? Stream.empty()
+                        : Stream.of(visaDetails.getLinkedAccountId()))
+            .distinct();
+
+    linkedAccountIdsToRefresh.forEach(
+        linkedAccountId -> {
+          var linkedAccount = linkedAccountService.getLinkedAccount(linkedAccountId);
+          try {
+            linkedAccount.ifPresentOrElse(
+                this::authAndRefreshPassport,
+                () -> log.info("No linked account found when trying to validate passport."));
+          } catch (Exception e) {
+            log.info("Failed to refresh passport, will try again at the next interval.", e);
+          }
+        });
+
+    return visaDetailsList.size();
+  }
+
+  /**
+   * Get a new passport for each linked accounts with visas or passports expiring within
+   * externalCredsConfig.getVisaAndPassportRefreshInterval time from now
+   *
+   * @return the number of linked accounts with expiring visas or passports
+   */
+  public int refreshExpiringPassports() {
+    var refreshInterval = externalCredsConfig.getVisaAndPassportRefreshDuration();
+    var expirationCutoff = new Timestamp(Instant.now().plus(refreshInterval).toEpochMilli());
+    var expiringLinkedAccounts =
+        linkedAccountService.getLinkedAccountsWithExpiringPassportsOrVisas(expirationCutoff);
+
+    for (LinkedAccount linkedAccount : expiringLinkedAccounts) {
+      try {
+        authAndRefreshPassport(linkedAccount);
+      } catch (Exception e) {
+        log.info("Failed to refresh passport, will try again at the next interval.", e);
+      }
+    }
+
+    return expiringLinkedAccounts.size();
+  }
+
+  public String getProviderAccessToken(
+      String userId, Provider provider, AuditLogEvent.Builder auditLogEventBuilder) {
+    var linkedAccount =
+        linkedAccountService
+            .getLinkedAccount(userId, provider)
+            .orElseThrow(
+                () ->
+                    new NotFoundException(
+                        String.format(
+                            "No linked account found for user ID: %s and provider: %s. "
+                                + "Please go to the Terra Profile page External Identities tab "
+                                + "to link your account for this provider.",
+                            userId, provider)));
+    if (linkedAccount.getExpires().before(Timestamp.from(Instant.now()))) {
+      throw new ForbiddenException(
+          String.format(
+              "The linked account found for user ID: %s and provider: %s has expired. "
+                  + "Please go to the Terra Profile page External Identities tab "
+                  + "to re-link your account for this provider.",
+              userId, provider));
+    }
+    var providerProperties = externalCredsConfig.getProviderProperties(provider);
+    return accessTokenCacheService.getLinkedAccountAccessToken(
+        linkedAccount, new HashSet<>(providerProperties.getScopes()), auditLogEventBuilder);
+  }
+
+  @VisibleForTesting
+  void authAndRefreshPassport(LinkedAccount linkedAccount) {
+    if (linkedAccount.getExpires().toInstant().isBefore(Instant.now())) {
+      invalidateLinkedAccountAndRemovePassport(linkedAccount);
+    } else {
+      try {
+        var linkedAccountWithRefreshedPassport = getRefreshedPassportsAndVisas(linkedAccount);
+        linkedAccountService.upsertLinkedAccountWithPassportAndVisas(
+            linkedAccountWithRefreshedPassport);
+        var transactionClaim =
+            linkedAccountWithRefreshedPassport
+                .getPassport()
+                .flatMap(p -> jwtUtils.getJwtTransactionClaim(p.getJwt()));
+        auditLogger.logEvent(
+            new AuditLogEvent.Builder()
+                .auditLogEventType(AuditLogEventType.LinkRefreshed)
+                .provider(linkedAccount.getProvider())
+                .userId(linkedAccount.getUserId())
+                .externalUserId(linkedAccount.getExternalUserId())
+                .transactionClaim(transactionClaim)
+                .build());
+
+      } catch (IllegalArgumentException iae) {
+        throw new ExternalCredsException(
+            String.format("Could not contact issuer for provider %s", linkedAccount.getProvider()),
+            iae);
+      } catch (OAuth2AuthorizationException oauthEx) {
+        // if it looks like the refresh token will never work, delete the passport
+        if (unrecoverableOAuth2ErrorCodes.contains(getRootOAuth2ErrorCode(oauthEx))) {
+          log.info(
+              String.format(
+                  "Caught unrecoverable oauth2 error code refreshing passport for user id [%s].",
+                  linkedAccount.getUserId()),
+              oauthEx);
+          if (linkedAccount.getId().isEmpty()) {
+            throw new ExternalCredsException("linked account id missing");
+          }
+          invalidateLinkedAccountAndRemovePassport(linkedAccount);
+        } else {
+          // log and try again later
+          throw new ExternalCredsException("Failed to refresh passport: ", oauthEx);
+        }
+      }
+    }
+  }
+
+  private LinkedAccountWithPassportAndVisas getRefreshedPassportsAndVisas(
+      LinkedAccount linkedAccount) {
+    var clientRegistration =
+        providerOAuthClientCache.getProviderClient(linkedAccount.getProvider());
+    var accessTokenResponse =
+        oAuth2Service.authorizeWithRefreshToken(
+            clientRegistration,
+            new OAuth2RefreshToken(linkedAccount.getRefreshToken(), null),
+            Collections.emptySet());
+
+    // save the linked account with the new refresh token and extracted passport
+    var linkedAccountWithRefreshToken =
+        Optional.ofNullable(accessTokenResponse.getRefreshToken())
+            .map(
+                refreshToken ->
+                    linkedAccountService.upsertLinkedAccount(
+                        linkedAccount.withRefreshToken(refreshToken.getTokenValue())))
+            .orElse(linkedAccount);
+
+    // update the passport and visas
+    var userInfo =
+        oAuth2Service.getUserInfo(clientRegistration, accessTokenResponse.getAccessToken());
+    return jwtUtils.enrichAccountWithPassportAndVisas(linkedAccountWithRefreshToken, userInfo);
+  }
+
+  @VisibleForTesting
+  boolean validateVisaWithProvider(VisaVerificationDetails visaDetails) {
+    var providerProperties = externalCredsConfig.getProviderProperties(visaDetails.getProvider());
+    if (providerProperties == null) {
+      throw new NotFoundException(
+          String.format("Provider %s not found", visaDetails.getProvider()));
+    }
+
+    var validationEndpoint =
+        providerProperties
+            .getValidationEndpoint()
+            .orElseThrow(
+                () ->
+                    new NotFoundException(
+                        String.format(
+                            "Validation endpoint for provider %s not found",
+                            visaDetails.getProvider())));
+
+    var response =
+        WebClient.create(validationEndpoint)
+            .get()
+            .uri(uriBuilder -> uriBuilder.queryParam("visa", visaDetails.getVisaJwt()).build())
+            .retrieve();
+    var responseBody =
+        response
+            .onStatus(HttpStatusCode::isError, clientResponse -> Mono.empty())
+            .bodyToMono(String.class)
+            .block(Duration.of(1000, ChronoUnit.MILLIS));
+
+    log.info(
+        "Got visa validation response.",
+        Map.of(
+            "linkedAccountId", visaDetails.getLinkedAccountId(),
+            "providerName", visaDetails.getProvider().toString(),
+            "validationResponse", Objects.requireNonNullElse(responseBody, "[null]")));
+
+    var visaValid = "valid".equalsIgnoreCase(responseBody);
+
+    if (visaValid) {
+      passportService.updateVisaLastValidated(visaDetails.getVisaId());
+    }
+    return visaValid;
   }
 }
